changed ft_atol to long long, might split them up eventually to long and long long

+++ ft_atol.c
@@ -6,17 +6,17 @@
 /*   By: ohaker <ohaker@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/05/24 19:42:26 by ohaker            #+#    #+#             */
-/*   Updated: 2025/09/23 19:49:35 by ohaker           ###   ########.fr       */
+/*   Updated: 2025/11/12 19:05:38 by ohaker           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-long	ft_atol(const char *str)
+long long	ft_atol(const char *str)
 {
 	int		x;
 	int		neg;
-	long	res;
+	long long	res;
 
 	x = 0;
 	neg = 1;
